fix greedyLayout with weighted input
There was a bug in applying weights for non-zero intersection sizes. The
greedy layout call wasn't considering the weights, leading to generating the
same initial layout as in the unweighted version. This frequently lead to
layouts that the optimizer couldn't step out of.

Fix this by considering weights when determining which set to lay out first in
the greedyLayout call