Replaced d3 src location with https-supported CDN.
Thanks so much for this library! I use a Chrome extension called HTTPS Everywhere (https://www.eff.org/https-everywhere) that automatically uses https when it is available. So that caused me to notice that the examples you linked to in the README won't work on Chrome when https is used because Chrome refuses to load d3.min.js because its URL is http. Since the d3 website doesn't support https, I swapped these links with a CDN that does support https. If you accept this pull request, don't forget to merge master into the gh-pages branch so the changes will propogate to the links in the README.