venn diagram normalization code
Different algorithms can produce different looking output that is equally
correct. This code attempts to remove the major differences by axis aligning
the major sets, taking the mirror image if needed, and re-arranging all the
disjoint clusters so that they are close to one another (in a rough grid
pattern).