Return stats on how the chart was laid out
Sometimes its nice to have some properties on how things were laid out: for
instance the textCentres were needed for issue #46, and it can also be useful
to have info on circles, transitions etc for later customization. Return those
from the chart call, and add an example of using those to add a sublabel
underneath the main label.